seperate out tests and add scripts
